requirements.txt
adb-shell==0.4.3
adbutils==1.0.6
alembic>=1.9.0
aniso8601==9.0.1
apkutils2==1.0.0
Appium-Python-Client==2.6.1
appnope==0.1.3
asn1crypto==0.24.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Authlib==1.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker>=1.6.2
boto3==1.26.44
botocore==1.29.44
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
cigam==0.0.3
click>=8.1.3
click-plugins==1.1.1
cryptography==3.2.1
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
dominate==2.3.1
EasyProcess==1.1
ecdsa==0.18.0
email-validator==1.3.1
entrypoint2==1.1
entrypoints==0.4
enum34==1.1.6
envs==1.4
eventlet==0.30.2
fastjsonschema==2.19.0
Flask==2.2.5
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate>=4.0
Flask-Moment>=1.0.0
Flask-Principal==0.4.0
flask-restx==0.5.1
Flask-Script==2.0.5
Flask-Security-Fork==2.0.1
Flask-SQLAlchemy>=3.0
Flask-SSE==1.0.0
Flask-Uploads==0.2.1
Flask-WTF>=1.2
fonttools==4.38.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
http-client==0.1.22
idna==2.6
infinity==1.5
intervals==0.9.2
ipaddress==1.0.19
ipykernel==6.16.2
ipython==7.34.0
itsdangerous>=2.1.2
jedi==0.19.1
Jinja2>=3.0,<3.2.0
jmespath==1.0.1
jsonify==0.5
jsonschema==2.6.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.5
lark-parser==0.12.0
lcov-cobertura==2.0.1
Mako==1.0.7
markup==0.2
MarkupSafe==2.0.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==0.8.4
mss==6.1.0
nbclient==0.5.6
nbconvert==5.6.1
nbformat==5.4.0
msoffcrypto-tool==5.0.0
nest-asyncio==1.5.8
openpyxl==3.1.2
flask-apscheduler==1.13.1
pandas==2.1.4
flask-reuploads==1.0.1.dev0
outcome==1.2.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
paypal==1.2.5
paypalrestsdk==1.13.1
pexpect==4.4.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2-binary==2.9.3
ptyprocess==0.5.2
public==0.0.46
py==1.11.0
pyasn1==0.4.8
pycparser==2.18
pycryptodome==3.3.1
pycurl==7.45.1
pyelftools==0.28
Pygments==2.17.2
PyMySQL==0.7.11
pyOpenSSL==17.5.0
pyparsing==3.0.9
pyscreenshot==3.0
python-dateutil==2.8.2
python-editor==1.0.3
python-jose-cryptodome==1.3.2
pytz==2023.3.post1
pyzmq==25.1.2
redis==4.3.3
requests==2.28.1
required==0.4.0
retry==0.9.2
rsa==4.9
runipy==0.1.5
s3transfer==0.6.0
selenium==4.4.3
six==1.11.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy>=2.0.0
SQLAlchemy-Utils==0.41.2
Strip==0.0.1.dev14
stripe==1.77.2
tcms-api==12.2
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
validators==0.20.0
visitor==0.1.3
warrant==0.6.1
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug>=2.3.7,<3.0
whichcraft==0.6.1
wincertstore==0.2
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
WTForms-SQLAlchemy==0.3
xmltodict==0.13.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
python-dotenv
